requirements.txt
fastapi==0.103.1
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
uvicorn==0.23.2
python-dotenv==1.0.0